comment module

+++ src/user/users.service.ts
@@ -1,12 +1,13 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
-import { Repository } from 'typeorm';
+import { IntegerType, Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
 import * as bcrypt from 'bcrypt';
 import { CompanyInvite } from '../company/entities/company-invite.entity';
 import { UserCompany } from '../company/entities/user-company.entity';
 import { Company } from '../company/entities/company.entity';
+import { UpdateUserDto } from './dto/update-user.dto';
 
 @Injectable()
 export class UsersService {
@@ -20,46 +21,76 @@ export class UsersService {
     private readonly companyRepository: Repository<Company>,
   ) {}
   async createUser(createUserDto: CreateUserDto): Promise<string> {
+    const usernameExists = await this.userRepository.exists({
+      where: { username: createUserDto.username },
+    });
+    if (usernameExists)
+      return 'Username already exists';
     const user: User = new User();
     user.username = createUserDto.username;
     user.firstname = createUserDto.firstname;
-    user.lastname = createUserDto.lastname;
+    user.lastname = createUserDto?.lastname;
     user.password = await this.hashString(createUserDto.password);
     await this.userRepository.save(user);
     return `Successfully created user with username "${createUserDto.username}".`;
   }
+  async updateUser(
+    updateUserDto: UpdateUserDto,
+    userId: number,
+  ): Promise<string> {
+    updateUserDto.password = await this.hashString(updateUserDto.password);
+    const result = await this.userRepository.update(userId, updateUserDto);
+    if (result.affected === 0) return 'no such record found with your user id';
+    return 'Updated Profile successfully.';
+  }
+
   async findOneById(id: number) {
     return await this.userRepository.findOne({
       select: { username: true, firstname: true },
       where: { id },
     });
   }
+
   async findOneByUsername(username: string): Promise<User | null> {
     return this.userRepository.findOne({ where: { username } });
   }
+
   async findAll(): Promise<User[]> {
     return this.userRepository.find({
       select: { id: true, username: true, firstname: true },
     });
   }
-  async deleteOne({
-    username,
-    password,
-  }: {
-    username: string;
-    password: string;
-  }): Promise<string> {
+
+  async deleteOne(userId: number, password: string): Promise<string> {
+    if (!password) {
+      throw new NotFoundException('Please provide password');
+    }
     const user: User | null = await this.userRepository.findOne({
-      where: { username },
+      where: { id: userId },
     });
-    if (!user) return 'no such user exists';
+    if (!user) return 'No such user exists.';
     const isMatch: boolean = await bcrypt.compare(password, user?.password);
     if (!isMatch) {
       return 'invalid user';
     }
     await this.userRepository.delete(user?.id);
-    return 'successfully deleted user';
+    return 'Successfully deleted user.';
   }
+  async findMyCompanies(userId: number) {
+    const user = await this.userRepository.findOne({ where: { id: userId } });
+    if (!user) return 'No such user exists.';
+    const companies = await this.userCompanyRepository
+      .createQueryBuilder('userCompany')
+      .leftJoinAndSelect('userCompany.companyId', 'company')
+      .select(['company.name', 'userCompany.role'])
+      .where('userCompany.userId=:userId', { userId })
+      .getMany();
+    if (companies) {
+      return companies;
+    }
+    return 'No companies found.';
+  }
+
   async findInvitationsToUser(userId: number): Promise<CompanyInvite[]> {
     const sendToUser = await this.userRepository.findOne({
       where: { id: userId },
@@ -117,9 +148,11 @@ export class UsersService {
       await this.companyInviteRepository.delete(invitationId);
 
       return 'Successfully accepted invitation';
-    } else {
+    } else if (acceptInvitation === 'reject') {
       await this.companyInviteRepository.delete(invitationId);
       return 'Successfully rejected invitation';
+    } else {
+      return 'you can only put "accept" or "reject" in url.';
     }
   }
